requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch>=2.6.0
torchvision>=0.21.0
diffusers>=0.24.0
Pillow>=10.0.0
matplotlib>=3.8.0
huggingface-hub>=0.19.0
transformers>=4.36.0
accelerate>=0.25.0 